adapter: rename `subscribe.rs` to `active_compute_sink.rs`
