+++ src/adapter/src/active_compute_sink.rs
